refactor(cli-ctx): relocate ctx object to parent module
